requirements.txt
pulsar-client==3.8.0
boto3==1.34.0
python-dateutil==2.8.2
requests==2.31.0
databricks-cli>=0.18.0
databricks-connect==16.4.*
jupyter>=1.0.0
ipykernel>=6.0.0 